fix: refactoring and change migration

+++ src/utils.py
@@ -54,12 +54,13 @@ async def lifespan(app: FastAPI):
 
 def clear_media_path():
     folder_path = os.path.join("static", "media")
-    for item in os.listdir(folder_path):
-        item_path = os.path.join(folder_path, item)
-        if os.path.isfile(item_path):
-            os.remove(item_path)
-        elif os.path.isdir(item_path):
-            shutil.rmtree(item_path)
+    if os.path.exists(folder_path):
+        for item in os.listdir(folder_path):
+            item_path = os.path.join(folder_path, item)
+            if os.path.isfile(item_path):
+                os.remove(item_path)
+            elif os.path.isdir(item_path):
+                shutil.rmtree(item_path)
 
 
 async def save_photo(
